refactor: restructure some files
