+++ src/use-base-model/index.ts
@@ -1,6 +1,6 @@
 // for Model functions
-export { useBaseModel } from './use-model_base'
-export { useFeathersModel } from './use-model_feathers'
+export { useBaseModel } from './use-base-model'
+export { useFeathersModel } from './use-feathers-model'
 export { useModelEvents } from './wrap-model_events'
 export { wrapModelBase } from './wrap-model_base'
 export { wrapModelFeathers } from './wrap-model_feathers'
@@ -8,6 +8,6 @@ export { wrapModelFeathers } from './wrap-model_feathers'
 // for model instances
 export { useModelInstance } from './use-model-instance'
 export { useInstanceDefaults } from './use-model-instance-defaults'
-export { useModelInstanceFeathers } from './use-model-instance-feathers'
+export { useFeathersInstance } from './use-feathers-instance'
 
 export * from './types'
